backend/requirements.txt
fastapi==0.115.5
uvicorn==0.32.0
mysql-connector-python==8.3.0
pydantic==2.10.3
pydantic-settings==2.6.1
python-dotenv==1.0.1
pytest==8.3.3
httpx==0.27.2
pytest-cov==6.0.0
